files restructuring and bug fixes
