+++ client/src/components/Interview/Interview.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from "react";
 import InterviewList from "./InterviewList";
-import { Interview } from "../types";
-import { useGetInterviewsQuery } from "../redux/ApiSlice/meetApi";
-import { useAppDispatch, useAppSelector } from "../redux/hooks";
-import { addAllInterviews } from "../redux/interviewsSlice";
-import Schedule from "./Schedule";
-import Loader from "./Loader";
+import { Interview } from "../../types";
+import { useGetInterviewsQuery } from "../../redux/ApiSlice/meetApi";
+import { useAppDispatch, useAppSelector } from "../../redux/hooks";
+import { addAllInterviews } from "../../redux/interviewsSlice";
+import Schedule from "../Schedule";
+import Loader from "../Loader";
 
 interface Tabs {
   id: number;
