+++ client/src/components/Interview/InterviewCard.tsx
@@ -1,9 +1,9 @@
-import LinkIcon from "../icons/LinkIcon";
-import TrashIcon from "../icons/TrashIcon";
-import { useAppDispatch } from "../redux/hooks";
-import { deleteInterview } from "../redux/interviewsSlice";
-import { useDeleteInterviewsMutation } from "../redux/ApiSlice/meetApi";
-import { Interview } from "../types";
+import LinkIcon from "../../icons/LinkIcon";
+import TrashIcon from "../../icons/TrashIcon";
+import { useAppDispatch } from "../../redux/hooks";
+import { deleteInterview } from "../../redux/interviewsSlice";
+import { useDeleteInterviewsMutation } from "../../redux/ApiSlice/meetApi";
+import { Interview } from "../../types";
 
 const InterviewCard = ({ data }: { data: Interview }) => {
   const [removeInterview] = useDeleteInterviewsMutation();
