+++ server-api/src/index.ts
@@ -1,6 +1,5 @@
 import dotenv from "dotenv";
 import app from "./app";
-import express from "express";
 dotenv.config();
 
 const port = process.env.PORT || 3000;
